:recycle: refacto question routes component
