+++ client/src/scenes/Question/Question.jsx
@@ -1,16 +1,26 @@
-import { Navigate, Route, Routes } from 'react-router-dom'
+import { useQuery } from '@apollo/client'
+import { Loading } from 'components'
+import { Navigate, Route, Routes, useParams } from 'react-router'
 
-import QuestionContainer from './Question.container'
-
-import Edit from './scenes/Edit'
-import Random from './scenes/Random'
+import { GET_NODE } from './queries'
+import Answer from './scenes/Answer/Answer'
+import Edit from './scenes/Edit/Edit'
+import Read from './scenes/Read/Read'
 
 const Question = () => {
+  const params = useParams()
+  const { data, loading } = useQuery(GET_NODE, {
+    variables: { id: params.slug?.split('-').at(-1) },
+    skip: !params.slug,
+  })
+
+  if (loading) return <Loading />
+
   return (
     <Routes>
-      <Route path="new" element={<Edit />} />
-      <Route path="random/:tag?" element={<Random />} />
-      <Route path=":slug/*" element={<QuestionContainer />} />
+      <Route path="/" element={<Read zNode={data.zNode} loading={loading} />} />
+      <Route path="/edit" element={<Edit zNode={data.zNode} loading={loading} />} />
+      <Route path="/answer" element={<Answer zNode={data.zNode} loading={loading} />} />
       <Route path="*" element={<Navigate to="/" />} />
     </Routes>
   )
