+++ client/src/scenes/Question/index.js
@@ -1,1 +1,1 @@
-export { default } from './Question'
+export { default } from './QuestionRoutes'
